Finish migration to Cash Pipeline for remaining functions and clear out now unused functions from core
