+++ pallets/cash/src/internal/mod.rs
@@ -3,12 +3,13 @@ pub mod balance_helpers;
 pub mod change_validators;
 pub mod events;
 pub mod exec_trx_request;
+pub mod extract;
 pub mod liquidate;
-pub mod liquidity;
 pub mod lock;
 pub mod miner;
 pub mod next_code;
 pub mod notices;
 pub mod set_yield_next;
 pub mod supply_cap;
+pub mod transfer;
 pub mod validate_trx;
